+++ honeybee_energy/material/opaque.py
@@ -57,7 +57,7 @@ class EnergyMaterial(_EnergyMaterialOpaqueBase):
     """
     __slots__ = ('_roughness', '_thickness', '_conductivity',
                  '_density', '_specific_heat', '_thermal_absorptance',
-                 '_solar_absorptance', '_visible_absorptance')
+                 '_solar_absorptance', '_visible_absorptance', '_user_data')
 
     def __init__(self, identifier, thickness, conductivity, density, specific_heat,
                  roughness='MediumRough', thermal_absorptance=0.9,
@@ -73,6 +73,7 @@ class EnergyMaterial(_EnergyMaterialOpaqueBase):
         self.solar_absorptance = solar_absorptance
         self.visible_absorptance = visible_absorptance
         self._locked = False
+        self._user_data = None
 
     @property
     def roughness(self):
@@ -202,6 +203,23 @@ class EnergyMaterial(_EnergyMaterialOpaqueBase):
         """The heat capacity per unit area of the material [kg/K-m2]."""
         return self.mass_area_density * self.specific_heat
 
+    @property
+    def user_data(self):
+        """Get or set an optional dictionary for additional meta data for this object.
+
+        This will be None until it has been set. All keys and values of this
+        dictionary should be of a standard Python type to ensure correct
+        serialization of the object to/from JSON (eg. str, float, int, list, dict)
+        """
+        return self._user_data
+    
+    @user_data.setter
+    def user_data(self, value):
+        if value is not None:
+            assert isinstance(value, dict), 'Expected dictionary for honeybee_energy' \
+                'object user_data. Got {}.'.format(type(value))
+        self._user_data = value
+
     @classmethod
     def from_idf(cls, idf_string):
         """Create an EnergyMaterial from an EnergyPlus text string.
@@ -256,7 +274,7 @@ class EnergyMaterial(_EnergyMaterialOpaqueBase):
         if 'display_name' in data and data['display_name'] is not None:
             new_mat.display_name = data['display_name']
         if 'user_data' in data and data['user_data'] is not None:
-            energymaterial.user_data = data['user_data']
+            new_mat.user_data = data['user_data']
 
         return new_mat
 
@@ -336,7 +354,7 @@ class EnergyMaterial(_EnergyMaterialOpaqueBase):
             self.specific_heat, self.roughness, self.thermal_absorptance,
             self.solar_absorptance, self._visible_absorptance)
         new_material._display_name = self._display_name
-        new_material._user_data = None if self.user_data is None else self.user_data.copy()
+        new_material._user_data = None if self._user_data is None else self._user_data.copy()
         return new_material
